chore: rename types to ft_types
