+++ freqtrade/exchange/exchange_utils.py
@@ -26,7 +26,7 @@ from freqtrade.exchange.common import (
     SUPPORTED_EXCHANGES,
 )
 from freqtrade.exchange.exchange_utils_timeframe import timeframe_to_minutes, timeframe_to_prev_date
-from freqtrade.types import ValidExchangesType
+from freqtrade.ft_types import ValidExchangesType
 from freqtrade.util import FtPrecise
