+++ freqtrade/ft_types/__init__.py
@@ -1,8 +1,8 @@
 # flake8: noqa: F401
-from freqtrade.types.backtest_result_type import (
+from freqtrade.ft_types.backtest_result_type import (
     BacktestHistoryEntryType,
     BacktestMetadataType,
     BacktestResultType,
     get_BacktestResultType_default,
 )
-from freqtrade.types.valid_exchanges_type import ValidExchangesType
+from freqtrade.ft_types.valid_exchanges_type import ValidExchangesType
