+++ freqtrade/ft_types/backtest_result_type.py
