+++ freqtrade/rpc/api_server/api_backtest.py
@@ -21,6 +21,7 @@ from freqtrade.data.btanalysis import (
 from freqtrade.enums import BacktestState
 from freqtrade.exceptions import ConfigurationError, DependencyException, OperationalException
 from freqtrade.exchange.common import remove_exchange_credentials
+from freqtrade.ft_types import get_BacktestResultType_default
 from freqtrade.misc import deep_merge_dicts, is_file_in_dir
 from freqtrade.rpc.api_server.api_schemas import (
     BacktestHistoryEntry,
@@ -32,7 +33,6 @@ from freqtrade.rpc.api_server.api_schemas import (
 from freqtrade.rpc.api_server.deps import get_config
 from freqtrade.rpc.api_server.webserver_bgwork import ApiBG
 from freqtrade.rpc.rpc import RPCException
-from freqtrade.types import get_BacktestResultType_default
 
 
 logger = logging.getLogger(__name__)
